Move wallet utils to core package
