+++ .changeset/chilly-balloons-agree.md
@@ -1,0 +1,5 @@
+---
+"@aptos-labs/wallet-adapter-core": minor
+---
+
+Added the following utility functions: `getAptosConnectWallets`, `partitionWallets`, `isInstalledOrLoadable`, `isInstallRequired`, and `truncateAddress`.
