+++ packages/wallet-adapter-core/src/utils/aptosConnect.ts
@@ -1,5 +1,6 @@
 import { WalletInfo } from "../LegacyWalletPlugins";
 import { AnyAptosWallet } from "../WalletCore";
+import { partitionWallets } from "./helpers";
 
 /** The base URL for all Aptos Connect wallets. */
 export const APTOS_CONNECT_BASE_URL = "https://aptosconnect.app";
@@ -11,4 +12,15 @@ export const APTOS_CONNECT_ACCOUNT_URL =
 /** Returns `true` if the provided wallet is an Aptos Connect wallet. */
 export function isAptosConnectWallet(wallet: WalletInfo | AnyAptosWallet) {
   return wallet.url.startsWith(APTOS_CONNECT_BASE_URL);
+}
+
+/**
+ * Partitions the `wallets` array so that Aptos Connect wallets are grouped separately from the rest.
+ * Aptos Connect is a web wallet that uses social login to create accounts on the blockchain.
+ */
+export function getAptosConnectWallets(wallets: ReadonlyArray<AnyAptosWallet>) {
+  const { defaultWallets, moreWallets } = partitionWallets(wallets, (wallet) =>
+    wallet.url.includes("aptosconnect.app")
+  );
+  return { aptosConnectWallets: defaultWallets, otherWallets: moreWallets };
 }
