+++ packages/wallet-adapter-core/src/utils/helpers.ts
@@ -8,8 +8,10 @@ import {
   SimpleEntryFunctionArgumentTypes,
 } from "@aptos-labs/ts-sdk";
 import { NetworkInfo as StandardNetworkInfo } from "@aptos-labs/wallet-standard";
-import { NetworkInfo } from "../LegacyWalletPlugins/types";
 import { convertNetwork } from "../LegacyWalletPlugins/conversion";
+import { NetworkInfo } from "../LegacyWalletPlugins/types";
+import { AnyAptosWallet } from "../WalletCore";
+import { WalletReadyState } from "../constants";
 
 export function isMobile(): boolean {
   return /Mobile|iP(hone|od|ad)|Android|BlackBerry|IEMobile|Kindle|NetFront|Silk-Accelerated|(hpw|web)OS|Fennec|Minimo|Opera M(obi|ini)|Blazer|Dolfin|Dolphin|Skyfire|Zune/i.test(
@@ -105,3 +107,48 @@ export const fetchDevnetChainId = async (): Promise<number> => {
   const aptos = new Aptos(); // default to devnet
   return await aptos.getChainId();
 };
+
+/**
+ * A function that will partition the provided wallets into two list — `defaultWallets` and `moreWallets`.
+ * By default, the wallets will be partitioned by whether or not they are installed or loadable.
+ * You can pass your own partition function if you wish to customize this behavior.
+ */
+export function partitionWallets(
+  wallets: ReadonlyArray<AnyAptosWallet>,
+  partitionFunction: (wallet: AnyAptosWallet) => boolean = isInstalledOrLoadable
+) {
+  const defaultWallets: Array<AnyAptosWallet> = [];
+  const moreWallets: Array<AnyAptosWallet> = [];
+
+  for (const wallet of wallets) {
+    if (partitionFunction(wallet)) defaultWallets.push(wallet);
+    else moreWallets.push(wallet);
+  }
+
+  return { defaultWallets, moreWallets };
+}
+
+/** Returns true if the wallet is installed or loadable. */
+export function isInstalledOrLoadable(wallet: AnyAptosWallet) {
+  return (
+    wallet.readyState === WalletReadyState.Installed ||
+    wallet.readyState === WalletReadyState.Loadable
+  );
+}
+
+/**
+ * Returns true if the user is on desktop and the provided wallet requires installation of a browser extension.
+ * This can be used to decide whether to show a "Connect" button or "Install" link in the UI.
+ */
+export function isInstallRequired(wallet: AnyAptosWallet) {
+  const isWalletReady = isInstalledOrLoadable(wallet);
+  const isMobile = !isWalletReady && isRedirectable();
+
+  return !isMobile && !isWalletReady;
+}
+
+/** Truncates the provided wallet address at the middle with an ellipsis. */
+export function truncateAddress(address: string | undefined) {
+  if (!address) return;
+  return `${address.slice(0, 6)}...${address.slice(-5)}`;
+}
