+++ packages/wallet-adapter-react/src/index.tsx
@@ -2,4 +2,3 @@ export * from "@aptos-labs/wallet-adapter-core";
 export * from "./WalletItem";
 export * from "./WalletProvider";
 export * from "./useWallet";
-export * from "./utils";
